+++ client/interfaces/components.ts
@@ -4,23 +4,37 @@ export interface StatelessPage<P = {}> extends React.SFC<P> {
   getInitialProps?: (ctx: any) => Promise<P>;
 }
 
-export interface IParameterChangesObject {
-  [key: string]: {
-    oldValue: any;
-    newValue: any;
-    type: any;
-    key: any;
-  };
+export interface IButton {
+  href?: string;
+  onClick?: any;
+  disabled?: boolean;
+  primary?: boolean;
+  large?: boolean;
+  type?:
+    | 'submit'
+    | 'default'
+    | 'primary'
+    | 'success'
+    | 'info'
+    | 'warning'
+    | 'danger'
+    | 'link'
+    | 'firstChoice'
+    | 'secondChoice';
+  active?: boolean;
+  width?: string;
+  color?: string;
+  bg?: string;
+  m?: string;
 }
 
-export interface IGovernanceSlateFormValues {
-  email: string;
-  title: string;
-  firstName: string;
-  lastName?: string;
-  organization?: string;
-  summary: string;
-  recommendation: string;
-  parameters: IParameterChangesObject;
-  stake: string;
+export interface ITag {
+  href?: string;
+  disabled?: boolean;
+  primary?: boolean;
+  large?: boolean;
+  type?: 'default' | 'primary' | 'success' | 'info' | 'warning' | 'danger' | 'link';
+  status?: string;
+  color?: string;
+  bg?: string;
 }
