+++ client/interfaces/eth.ts
@@ -1,0 +1,8 @@
+import { BasicToken, Gatekeeper, TokenCapacitor, ParameterStore } from '../types';
+
+export interface IContracts {
+  tokenCapacitor: TokenCapacitor;
+  gatekeeper: Gatekeeper;
+  token: BasicToken;
+  parameterStore: ParameterStore;
+}
