--- client/interfaces/ui.ts
@@ -1,34 +1,0 @@
-export interface IButton {
-  href?: string;
-  onClick?: any;
-  disabled?: boolean;
-  primary?: boolean;
-  large?: boolean;
-  type?:
-    | 'submit'
-    | 'default'
-    | 'primary'
-    | 'success'
-    | 'info'
-    | 'warning'
-    | 'danger'
-    | 'link'
-    | 'firstChoice'
-    | 'secondChoice';
-  active?: boolean;
-  width?: string;
-  color?: string;
-  bg?: string;
-  m?: string;
-}
-
-export interface ITag {
-  href?: string;
-  disabled?: boolean;
-  primary?: boolean;
-  large?: boolean;
-  type?: 'default' | 'primary' | 'success' | 'info' | 'warning' | 'danger' | 'link';
-  status?: string;
-  color?: string;
-  bg?: string;
-}
